+++ main/CMakeLists.txt
@@ -1,11 +1,10 @@
 idf_component_register(SRCS "main.c"
                     INCLUDE_DIRS ".")
 
-add_subdirectory(helpers)
-target_link_libraries(${COMPONENT_LIB} PUBLIC helpers)
-
-add_subdirectory(i2c_devices)
-target_link_libraries(${COMPONENT_LIB} PUBLIC i2c_devices)
+add_subdirectory(i2c)
+target_link_libraries(${COMPONENT_LIB} PUBLIC bmp280)
+target_link_libraries(${COMPONENT_LIB} PUBLIC cat24c256)
+target_link_libraries(${COMPONENT_LIB} PUBLIC mpu6050)
 
 add_subdirectory(control)
 target_link_libraries(${COMPONENT_LIB} PUBLIC control)
